add HOC and Login components
